Add: kconfig frontends ebuild
Kconfig is the configuration language used by the Linux kernel. This
package is a simple copy of the frontends and the parser found in the
Linux kernel source tree, with very minor changes to adapt them to being
built out of the kernel build infrastructure.

The kconfig-frontends are provided and maintained by
"Yann E. MORIN" <yann.morin.1998@free.fr>
and may be found at
http://ymorin.is-a-geek.org/projects/kconfig-frontends
The sources are located at
http://ymorin.is-a-geek.org/hg/kconfig-frontends

Signed-off-by: Daniel Kriesten <krid@tu-chemnitz.eu>